refactor: handleBlockClick, handleBlockInput, handleKeyDown을 useBlockOperation 커스텀훅으로 분리
